sc_guest: Support CreateEPTView for EPT container
The most significant work for creating EPT container is to isolate
its memory in a separate EPT table. This patch is to deal with this
work. It will call KVM interface to create a new EPT table with GPA
of the first executed address of container init process. After that,
when the container init process is being executed, KVM will know to
switch view to this new EPT table rather than guest EPT table.

Signed-off-by: Chunyan Liu <liuchunyan9@huawei.com>